Fix memory leak issue
Fix #22
